fix echo, makefile

+++ includes/slice.h
@@ -1,8 +1,7 @@
 #include "mem.h"
-
 typedef struct
 {
-    int cap;
-    int len;
+    size_t cap;
+    size_t len;
     byte_t *pool;
 } slice_t;
